Fix apport-kde crashing
Make sip throw away all objects instead of calling dtors for each
as this causes apport to crash while exiting. This is just a workaround
and also introduces another bug which causes apport to launch a dialog
saying that it could launch the browser, when in reality it did. But not
crashing is prioritized over a single annoying dialog.