Use pr.extract() method to write big binaries to disk directly
Use the new Problem.Report.extract() method to write the large 
binary element to disk without loading them first. A first pass
with binary=False is done to identify the binary element and
write the text-only element to disk.

A second pass is done and ProblemReport.extract() is used on each
binary element to write those to disk directly.