apport_python_hook.py: Only run for supported versions
Apport's Python crash handler runs for every installed version of Python
e.g. on Ubuntu 18.04 LTS it will run for crashes with python3.6 or
python3.7 as an interpreter. However, Python apt modules are only
available for the supported version of Python so the crash handler
generates an exception which is annoying.

If python apt modules are not built for the Python version then do
capture the crash.

Import change from apport 2.20.11-0ubuntu34.

LP: #1774843