apport: error out when -w option is used on wayland
The -w,--window command line flag results in apport using xprop to get
the window's PID. On wayland, xprop cannot be used, and trying to call
xprop results in a non-responsive apport UI. Due to wayland's
architecture, there is no viable replacement for xprop. When the -w
option is used on wayland, error out with a helpful message.

LP: #1952947