whoopsie-upload-all: Do not upload after EOFError/zlib.error
apport 2.20.10-0ubuntu27 changed the error handling for the
`add_gdb_info` call. This change has the undesired side-effect that the
report processing will continue also for errors that do not have an
`errno` attribute (like `EOFError` and `zlib.error`).

So also return from the report processing if the error has no `errno`
attribute to restore the previous behavior.

Fixes: 617f8c325ed2 ("merge changes which made it into the Ubuntu package but not upstream")
Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>