Drop debug argument from ParseSegv.__init__
The test case `test_debug` changes `sys.stderr`, but does not reset it
after the test finishes. That leaves stderr changed for the following
test cases (which is not desired).

Since only this one test case specifies the `debug` argument, drop this
argument and also the logging setup code (which would influence the
following test cases as well). Use `assertLogs` in `test_debug` to
change the logging level and also check that something is logged on
debug level.

Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>