tests/ui: Generate test crash files directly
The integration tests in `test_ui.py` need a crash report with real
crash data, but they do not care how this crash report is generated.
`_gen_test_crash` is used for creating the crash reports. It relies on a
properly setup apport that will create crash dump files for crashed
programs. Apport is not configured during Debian package build and
therefore creating the crash report will fail in that environment.

Generate the test crash files directly using `gdb`. This will work
regardless of apport being configured as crash handler in the system.

Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>