Grab a slice of JournalErrors around the crash time
For gnome-shell the desktop team wants the journal errors in the seconds
after the crash to see the Javascript stack dump (if any). But for some
other crashes they need to also see what happened just before the crash.

The crash data collection can happen after the crash, not immediately at
the time of the crash. This will lead to the `JournalErrors` entries to
show the logs seconds after the crash.

Grab a slice of JournalErrors around the crash time (if available)
instead of 1000 lines.

Bug: https://launchpad.net/bugs/1962454
Restructed-by: Benjamin Drung <bdrung@ubuntu.com>