tests: Fix race in tests for run_report_bug()
The two test cases that run `run_report_bug` need to wait until
`GTKUserInterface.ui_present_report_details` reaches `Gtk.main()`.
Otherwise the click event might be emitted before the event is
connected.

Instead of waiting for the send button to be visible, wait for the send
button to be connected (to `dialog_crash_dismissed`).

Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>