tests: Generate test crash file for test_add_gdb_info_script directly
The integration test `test_add_gdb_info_script` needs a crash report
with real crash data, but it does not care how this crash report is
generated. It relies on a properly setup Apport that will create the
crash dump file for crashed programs. Apport is not configured during
Debian package build and therefore creating the crash report will fail
in that environment.

Generate the test crash file directly using `gdb`. This will work
regardless of Apport being configured as crash handler in the system.

Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>