tests: Fix test_kerneloops_nodetails if kernel is not installed
The kernel package might not be installed in the test environment (e.g.
in a schroot or Docker container). Then `test_kerneloops_nodetails` will
fail:

```
FAIL: test_kerneloops_nodetails (tests.test_ui_gtk.T)
Kernel oops report without showing details
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/usr/lib/python3.10/unittest/mock.py", line 1369, in patched
    return func(*newargs, **newkeywargs)
  File "tests/test_ui_gtk.py", line 820, in test_kerneloops_nodetails
    self.assertTrue('Dependencies' in r)
AssertionError: False is not true
```

Check if the package is not installed and skip checking the presence of
`Dependencies` in the report in this case.

Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>