tests: Fix test_add_gdb_info_script on armhf
GDB on armhf needs `bash-dbgsym` installed for generating a proper stack
trace for the `test_add_gdb_info_script` test case. If `bash-dbgsym` is
not installed, the stack unwinder will guess the stack frame boundaries
and this guess will be wrong. See
[gdb reports 'corrupt stack' on armhf without symbols](https://launchpad.net/bugs/1325503).

Since the installation of `bash-dbgsym` is part of the environment, the
test suite should not try to install it. Instead change the
`test_add_gdb_info_script` test case to check for parts in the stack
trace that will be always present.

Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>