Add support for qastaging.launchpad.net
There are two test environments, `staging.launchpad.net` and
`qastaging.launchpad.net` which contain a copy of the Launchpad data and
newer code. The staging server additionally has any database changes
which the Launchpad developers are testing out on in.

Add support for `qastaging.launchpad.net` and default the tests to use
it, because it is probably more stable than `staging.launchpad.net`.

Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>