Suggest installing python3-launchpadlib if missing and needed
A user reported a bug that apport-collect does not work, because it
fails with the error message: "The python3-launchpadlib package is not
installed. This functionality is not available."

Clarify the error message by telling the user to install the
`python3-launchpadlib` package. The package is called
`python3-launchpadlib` on Fedora and on all Debian based distributions.

Bug: https://launchpad.net/bugs/1958059
Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>