apport-kde: Drop old workaround for bug in SIP destructor
apport-kde fails if the Python module `sip` is not installed:

```
Traceback (most recent call last):
  File "/usr/share/apport/apport-kde", line 21, in <module>
    import sip
ModuleNotFoundError: No module named 'sip'
```

The workaround for a crash with SIGSEGV in
QWidgetPrivate::deleteTLSysExtra() by skipping the destructors of SIP
was introduced in 2014. Assume that the underlying bug was fixed since
then and just drop the workaround.

Bug: https://launchpad.net/bugs/1980553
Signed-off-by: Benjamin Drung <bdrung@ubuntu.com>