CI: Set environment DEBIAN_FRONTEND=noninteractive
Installing the dependencies on Ubuntu 22.10 (kinetic) hangs:

```
Configuring libdebuginfod-common
--------------------------------

While debugging programs (with GDB, for example) or using debuginfo-consumer
applications, it is possible to connect to Ubuntu's debuginfod server and
download the necessary debug information for the program you are debugging
on-the-fly, without the need to configure the ddebs apt repository nor
installing any ddeb package.  This service is maintained by Ubuntu, and the
only information you will have to send to it is the Build-ID hash of the
program(s)/library(ies) being debugged.

Connect to Ubuntu's debuginfod server to download debug symbols? [yes/no]
```

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>