Disable debuginfod when collecting a report
With the new Ubuntu debuginfod
service (https://debuginfod.ubuntu.com), and with the prospect of
having the system automatically fetch debuginfo from the internet
without user intervention, it is necessary to adjust apport to cope
with this scenario.

I had a conversation with bdmurray and he was concerned that having a
debuginfod-enabled GDB generate the corefiles that are eventually
submitted to Ubuntu can be a problem.

This commit disables GDB's and valgrind's debuginfod support when
collecting a report.  I chose not to touch GDB's settings on
apport-retrace because I think we should honour the user choice there.

Bug: https://launchpad.net/bugs/1989803
Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>