test: Test wait_for_core_file() helper method
The `wait_for_core_file` helper method can success on the first try.
Then the "sleep and retry" part of the function is not used which leads
to a reduced code coverage.

Write test cases for the `wait_for_core_file` helper method to have full
code coverage for it.

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>