fix: Ensure full coverage of test_get_file_package_uninstalled
The code coverage for `test_get_file_package_uninstalled` is not 100%
because only one path of the `"updates" in cache_dir_files[0]`
if-condition is taken. `os.listdir` returns a list of files in arbitrary
order and therefore the entries in `cache_dir_files` are not stable.
Therefore the coverage changes from one to the other branch from time to
time.

Sort the output of `os.listdir` and replace the if-condition with a list
comprehension for 100% code coverage.

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>