fix: Cover bpthread.is_alive() while loop
The `REThread` calling `search_bug_patterns` can finish quickly so that
`bpthread.is_alive()` in `collect_info` returns `False` on the call.
Then the while loop not entered which leads to a reduced code coverage.

Enhance the test case `test_collect_info_package` to mock
`search_bug_patterns` to sleep until the while loop is entered (to have
code coverage for it).

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>