ui: upload: new API to display messages during upload
The backend might need to prompt the user to do stuff during the upload,
e.g. log into an external service. This API will be needed for the
upcoming Github backend.

Since the upload is on a separate thread, it's fairly dangerous to have
the callback directly interact with the UI object. Instead, we use a
simple queue that the UI thread will be polling, with Events to signal
back to the upload thread that the dialog has been closed.

Note that it would *probably* work fine without that because of the GIL,
and it could be that our UIs are thread-safe anyway, but I'd prefer not
to bet on those facts.

V2:
* formatting
* fold the get_nowait() + thread.join() into a blocking get()
* add missing changes on crashdb implementations. Whoops!

V3:
* Fix tests for bisectability

V4:
* Squash feature and tests into a single commit

Co-authored-by: Nathan Pratta Teodosio <nathan.teodosio@canonical.com>
Co-authored-by: Edu Gómez Escandell <edu.gomez.escandell@canonical.com>
Signed-off-by: Simon Chopin <simon.chopin@canonical.com>