feat: Already add package info during crash
The package info is added to the crash report when the report is
reported by the user or when it is uploaded. There can be quite some
time delay and the list of installed packages could have change since
then. There are crashes on errors.ubuntu.com where the reported version
is newer than the code in the Python stack trace. So this kind of
discrepancy happens in the real world.

So try to add the package info during crash. Add a `_HooksRun` entry to
the report to document that the hooks were not run (for
`whoopsie-upload-all`). Change `add_package_info` to not overwrite
existing entries.

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>