tests: skip the GH upload test if internet isn't available
Typically, this would be the case when building a Debian or Ubuntu
package.

Another approach would have been to mock out completely the function
that reaches out to GH, but I figure it's a good idea to have it
actually run regularly to ensure we quickly detect issues such as
expiring tokens.

Signed-off-by: Simon Chopin <simon.chopin@canonical.com>