feat: allow recent_syslog to collect all journal messages that the user can see
When using journalctl to extract the "recent" syslog messages, the
--system option was unconditionally passed. This means messages from the
user journal are not included. Added a parameter of recent_syslog to
/not/ pass the --system option to journalctl ; so that everything the
current user sees (including output of commands run with systemd-run
--user) is included if requested.

To use the new option, the caller must invoke recent_syslog with the
journald_system_only=False keyword argument. By default, its value is
set to True so that the default behavior remains unchanged.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>