apt_dpkg: sandbox: pivot around the apt conf dir rather than source.list
In the brave new world of deb822, there might not *be* a source.list
file that we can rely on!!

Note that this drops an early check for the existence of the source.list
file. That check was of fairly low value in terms of validating input in
the first place (source.list could be empty or full of garbage), and
replacing it with the apt dir would even lower that value, as the dir
itself could be empty. In addition to that, the exception raised is too
generic to be truly useful.