crash-digger: Add support for using a gdb-sandbox
I'm working on moving the service which retraces apport crashes in
Launchpad to a different server and noticed that while the existing
service uses crash-digger it does not take advantage of the GDB sandbox
option which I added to apport some time ago. I've added an option to
crash-digger, --gdb-sandbox, which is then is then passed through as
argument to apport-retrace.