run-linters: Add --errors-only mode
Running the linters during the Ubuntu package build or in the
autopkgtest is not wanted because newer linter versions can introduce
new checks that can cause the build/autopkgtest to fail.

So add an `--errors-only` mode that skips all formatting checks, but
keeps the checks that can detect real errors. So run pylint with
`--errors-only` and mypy.

Bug: https://launchpad.net/bugs/2028881
Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>