tests: ui: fix crash file generation to also contain Dependencies field
We started collecting deps in the initial crash file to hopefully have a
more exact picture of the system at the time of the crash. This should
also be reflected in the test data.

This enables coverage for the commit
"ui: don't rely on Dependencies to know if collect_info() has been run"

Signed-off-by: Simon Chopin <simon.chopin@canonical.com>