launchpad: catch import error if httplib2 is missing
The launchpadlib Python modules uses and depends on the httplib2 module.
If launchpadlib is not installed, httplib2 might not be installed.

Move importing httplib2 to the guarded launchpadlib import which sets
`Launchpad` to `None` in an import failure case. Then the user will get
an error message presented on usage.

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>