problem_report: separate base64 encoding and gzip compression
Refactor _write_key_and_binary_value_compressed_and_encoded_to_file into
a generator taking in all sorts of data and compressing and splitting it
into byte chunks, and a function encoding an iterable of binary chunks
into an RFC822 base64 field.

The compression generator is still in need of further refactoring, but
it's still a step in the right direction.

Signed-off-by: Simon Chopin <simon.chopin@canonical.com>