problem_report: refactor logic of load and extract_keys
Rewrite the logic for `ProblemReport.load` and
`ProblemReport.extract_keys` to be more readable and extensible. Use
`_EntryParser` to get an iterator over the entries. Introduce more
iterator for decoding the base64 text or the plain text. Using iterators
allow using the same code in `ProblemReport.load` and
`ProblemReport.extract_keys`.

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>