refactor(apport): introduce process_crash
Carve out the parts of `process_crash_from_kernel_with_proc_pid` that
will be used by the systemd-coredump handler into the separate function
`process_crash`.

To avoid passing more parameters to `process_crash`, derive the needed
information from the report: The signal number `options.signal_number`
was stored as `Signal` in the report and can be used from there. Store
the PID in the report.

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>