fix: use context manager when manipulating GzipFiles (LP: #2051512)
Those objects are technically file-like objects, with the associated
semantics. In particular, starting with Python 3.12 the class using
internal buffering, which means write() calls won't necessarily directly
be reflected in the underlying storage object.

For the sake of consistency, the context manager approach has been
generalized to read() calls as well, except when we could get away with
removing the use of GzipFile altogether.