ui: fix parameter new for webbrowser.open
The `new` parameter in `webbrowser.open` takes an integer and not a
boolean: "If `new` is 0, the url is opened in the same browser window if
possible. If `new` is 1, a new browser window is opened if possible. If
`new` is 2, a new browser page (“tab”) is opened if possible."

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>