apport: Ignore container crash reports from systemd-coredump
systemd-coredump stores crashes that happen inside a container on the
host in case it cannot forward the crash to the container. Apport cannot
do anything useful with these crash reports, because the information
about the crashed process does not contain the process number seen from
inside the container and because the crashed process is already gone
when Apport gets the crash from systemd-coredump.

So let Apport just ignore crash reports from systemd-coredump if the
crashes happened inside containers.

Bug-Ubuntu: https://launchpad.net/bugs/2063349
Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>