tests: apport_retrace: be more flexible on stack trace format
The format of the stack trace can vary a bit depending on architecture,
optimization level, etc. The issue that triggered the investigation was
that on s390x, the instruction that triggers the exception isn't the
*first* instruction for the specific line in the source code, so the
precise insn offset is added to the frame line.

(cherry picked from commit 7d53a1f242905e530d18b34e5dfe126a69478c2e)