launchpad: don't systematically order retrace on private reports
Just because a report is marked as LaunchpadPrivate and has an actual
architecture attached to it doesn't mean it should be retraced. A
contrario, any report that is designated as LaunchpadPrivate should be
marked as private, even if get_os_info() was never called on it.

Fixes https://bugs.launchpad.net/ubuntu/+source/apport/+bug/2068933