packaging: let get_uninstalled_package always return str
`PackageInfo.get_uninstalled_package` is only used in the test suite.
The tests that use `get_uninstalled_package` expect it to return a valid
name. The code path that returns `None` is currently not hit.

Let `PackageInfo.get_uninstalled_package` always return a package and
raise a `ValueError` in case no uninstalled package was found. Otherwise
the callers would need to check for the possible `None` result.

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>