apt_dpkg: stop support passing a filename to _check_files_md5
Only `get_modified_files` calls `_check_files_md5` and passes the
content of a md5sum file to it. Support for specifying a md5sum file as
parameter is only used the the integration test.

Drop support for passing a filename to `_check_files_md5` to make the
code and type hints simpler.