test: wait_for_gdb_sleeping_child_process fix
Modify the `wait_for_gdb_sleeping_child_process` utility function
to make sure to wait until the test process is actually sleeping.
The expected behavior is that the process has already entered the
"sleeping" state when returning from this function. Some callers
use this function to generate a process and then examine the stack
trace at a certain state. However, there are steps between "tracing-stop"
and "sleeping", such as "running", that may cause the stack trace from
the inspected process to look slightly different when compared a stack
trace from when the process once has entered the "sleeping" state. This
can be seen on armhf in LP: #2073933.

This commit also generally cleans up the test case and adjusts to catch
the failing case from before.

Co-authored-by: Benjamin Drung <benjamin.drung@canonical.com>