test: fix flaky test_crash_system_slice
The `test_crash_system_slice` test is flaky due to the
`create_test_process` function sometimes returning the sleep
process before it has entered the sleeping state. This is
especially true when starting it indirectly using systemd-run
since systemd-run will return when the process has been execve()ed
but not quite started yet. This was fairly reproducible on armhf
(LP: #2076186) and easily reproducible on amd64 after removing the
`time.sleep` for "setup" time in `create_test_process`. Now the test
will look for the sleep process in a 5 second timeout loop and
fail with a more descriptive error about not finding the process.
Also utilize the `wait_for_sleeping_state` utility to wait to
properly wait until the sleep command has fully entered the "sleeping"
state before killing it.