test: wait_for_process_to_appear
Introduce helper method `wait_for_process_to_appear` which expects
the path of an executable, a set of known PIDs for that process, and
a timeout to wait for the PID to appear. If the PID is not found before
`timeout` seconds, a `TimeoutError` is raised. Otherwise, the PID
is returned as an `int`.