test: do not check for exact encoded gzip data in test_write_file
The test case `test_write_file` fails on s390x from time to time due to
the encoded chunks have a different on-disk representation. Instead of
the expected one compressed chunk

```
File: base64
 H4sICAAAAAAC/0ZpbGUA
 c3RyxIAMcBAFAK/2p9MfAAAA
```

there could be two compressed chunks:

```
File: base64
 H4sICAAAAAAC/0ZpbGUA
 cnTChAxwEA==
 BRgAr/an0x8AAAA=
```

The raw data is identical, but the zlib compressed data on disk differs!
So make the test case less picky by just checking that the data is
compressed and that the decompressed data is correct.

Bug: https://launchpad.net/bugs/2076269