apt_dpkg: Replace apt-key with direct key download
apt-key has been removed in apt 2.9.17. Fetch the signing key
directly from launchpad and store it in a .asc file.

Change the test cases to compare the ASCII-armored keys rather than
check gpg output.

Also the unnamed test case was checking for an empty output as it
was broken a couple years ago. We'll fix that properly in a future
commit to actually use an unnamed PPA again.

moo

Fixes: a1f024f4fc17 ("test: use packages from the apport-hackers PPA")