problem_report: replace hasattr(x, "find") by isinstance(x, bytes | str)
The ProblemReport values must be of type `ProblemReportValue` which is a
union of different types: The types `bytes` and `str` have the attribute
`find` but the types `CompressedFile`, `CompressedValue`, and `tuple` do
not have the attribute `find`.

Replace `hasattr(x, "find")` by equivalent `isinstance` checks to make
it obvious what types are expected there.

The method `_write_binary_item_compressed_and_encoded` is only called
for binary keys. This method is the only caller of
`_generate_compressed_chunks`. Thus the value in
`_generate_compressed_chunks` cannot be of type `str`.