problem_report: write CompressedValue in chunks as well
In problem reports binary data is compressed in chunks of 1 MB. Each
compressed chunk is base64 encoded and written as one line in the
problem report file. This is the rule for all binary data formats except
for `CompressedValue`.

The problem report files are text files. Text editors like gedit might
choke on text files that have huge lines. So the problem report should
wrap the lines at a reasonable length.

So write `CompressedValue` in chunks as well.