ui: update_report: fix source/binary confusion
Rather than relying on binary packages being named the same as their
source package, we actually split the two of them. There's an underlying
assumption that we're reporting to a backend that cares about source
packages, so that's what we're basing it all on, by adding a field to
list the binary packages actually installed from that source package.

It might not work perfectly (the Package field will be marked with "not
installed" for glibc, for instance), but it's still much better than the
current state.

Bug-Ubuntu: https://bugs.launchpad.net/ubuntu/+source/apport/+bug/2102147