apport: use opened /proc/<pid> everywhere
Qualys discovered a vulnerability in apport (Ubuntu's core-dump handler),
and a similar vulnerability in systemd-coredump (which is the default
core-dump handler on Red Hat Enterprise Linux 9 and Fedora for example):
a race condition that allows a local attacker to crash a SUID program
and gain read access to the resulting core dump (by quickly replacing
the crashed SUID process with another process, before its /proc/pid/
files are analyzed by the vulnerable core-dump handler).

Do all accesses to files in `/proc/<pid>` through the `ProcPid` object
that opened the `/proc/<pid>` once.

CVE-2025-5054
Bug: https://launchpad.net/bugs/2107472