apport: do not forward crash for dump_mode == 2
Qualys discovered a vulnerability in apport (Ubuntu's core-dump handler),
and a similar vulnerability in systemd-coredump (which is the default
core-dump handler on Red Hat Enterprise Linux 9 and Fedora for example):
a race condition that allows a local attacker to crash a SUID program
and gain read access to the resulting core dump (by quickly replacing
the crashed SUID process with another process, before its /proc/pid/
files are analyzed by the vulnerable core-dump handler).

An attacker can kill-and-replace the crashed process before Apport
starts, but with the right timing it is also possible to give the kernel
enough time to write the core dump to the pipe (which allows the
attacker to read the memory of the crashed SUID process).

Do not forward crashes with dump mode of 2 to containers to prevent
attackers.

CVE-2025-5054
Bug: https://launchpad.net/bugs/2107472

(cherry picked from commit 907a2b7f9d30d77d8bf8db23193a469ce2a1923b)