apport: determine report owner by dump_mode
Qualys discovered a vulnerability in apport (Ubuntu's core-dump handler),
and a similar vulnerability in systemd-coredump (which is the default
core-dump handler on Red Hat Enterprise Linux 9 and Fedora for example):
a race condition that allows a local attacker to crash a SUID program
and gain read access to the resulting core dump (by quickly replacing
the crashed SUID process with another process, before its /proc/pid/
files are analyzed by the vulnerable core-dump handler).

Currently Apport decides who the owner of a /var/crash report should be
by looking up the owner of /proc/pid/stat (in
process_crash_from_kernel_with_proc_pid()), but, again, an attacker can
replace the originally crashed SUID process with a non-SUID process
(whose /proc/pid/stat belongs to the attacker) and therefore become the
owner of the crash report and gain access to the SUID core dump it
contains.

To fix that security issue: if the dump mode is 2 ("suidsafe"), make
sure that only root has access to the crash report and core dump.

CVE-2025-5054
Bug: https://launchpad.net/bugs/2107472