apport: support pidfd (%F) parameter from kernel
Qualys discovered a vulnerability in apport (Ubuntu's core-dump handler),
and a similar vulnerability in systemd-coredump (which is the default
core-dump handler on Red Hat Enterprise Linux 9 and Fedora for example):
a race condition that allows a local attacker to crash a SUID program
and gain read access to the resulting core dump (by quickly replacing
the crashed SUID process with another process, before its /proc/pid/
files are analyzed by the vulnerable core-dump handler).

An attacker can kill-and-replace the crashed process before Apport
starts, but with the right timing it is also possible to give the kernel
enough time to write the core dump to the pipe (which allows the
attacker to read the memory of the crashed SUID process).

To avoid this race condition, rely on the kernel to provide the pidfd
via `%F`. Apport can open `/proc/<pid>` and then check that the process
ID has not been reused.

CVE-2025-5054
Bug: https://launchpad.net/bugs/2107472